moved prepprojection in to maputil for parsing wmts options (#279)
